Shut up compiler warnings.

+++ src/m_cheat.h
@@ -47,12 +47,12 @@ typedef struct
     // settings for this cheat
 
     char sequence[MAX_CHEAT_LEN];
-    int sequence_len;
+    size_t sequence_len;
     int parameter_chars;
 
     // state used during the game
 
-    int chars_read;
+    size_t chars_read;
     int param_chars_read;
     char parameter_buf[MAX_CHEAT_PARAMS];
 } cheatseq_t;
